\name{createStateMatrix}
\alias{createStateMatrix}
\title{Create state matrix (one row per time holding bubble for each mole rat...}
\usage{createStateMatrix(x)
}
\description{Create state matrix (one row per time holding bubble for each mole rat}
\value{a multivariate zoo object}
\arguments{\item{x}{list of zoo objects from createZooObjects}
}

