\name{createZooObjects}
\alias{createZooObjects}
\title{Create zoo objects per mole rat from data frame...}
\usage{createZooObjects(x)
}
\description{Create zoo objects per mole rat from data frame}
\arguments{\item{matrix}{produced by readNMRData}
\item{list}{of zoo objects, one for reach ID in data set}
}

